Update Issue#id to long, bump version

+++ src/main/java/com/spotify/github/v3/issues/Issue.java
@@ -42,7 +42,7 @@ public interface Issue extends CloseTracking {
 
   /** ID. */
   @Nullable
-  Integer id();
+  Long id();
 
   /** URL. */
   @Nullable
